Added the measure and measurements tests to the full battery
