+++ tests/commands/__init__.py
@@ -1,4 +1,11 @@
+from .measure import TestMeasureCommand
+from .measurements import TestMeasurementsCommand
 from .probes import TestProbesCommand
 from .report import TestReportCommand
 
-__all__ = [TestProbesCommand, TestReportCommand]
+__all__ = [
+    TestMeasureCommand,
+    TestMeasurementsCommand,
+    TestProbesCommand,
+    TestReportCommand
+]
